torchvision/requirements.txt
numpy
matplotlib
pandas
Pillow
torch
torchvision
opencv-python